pyproject.toml
[project]
name = "stock-checker"
version = "0.1.0"
description = "CLI stock market analysis tool with charts and technical indicators"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "yfinance>=0.2.40",
    "matplotlib>=3.9",
    "mplfinance>=0.12.9",
    "numpy>=1.26",
    "flask>=3.0",
    "gunicorn>=22.0",
    "flask-sqlalchemy>=3.1",
    "cachetools>=5.3",
    "scipy>=1.12",
    "fpdf2>=2.7",
    "anthropic>=0.40",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.12",
    "beautifulsoup4>=4.12",
    "tenacity>=8.2",
    "PyYAML>=6.0",
    "pyarrow>=15.0",
    "requests>=2.31",
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[project.scripts]
stock-checker = "stock_checker.cli:main"
stock-checker-web = "stock_checker.app:run"
idx-anomaly = "stock_checker.idx_anomaly.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stock_checker"]
